Store resolved resources in the context.
